requirements.txt
Flask
Flask-cors
requests
python-dotenv